Resolve ZIP change detection bug
Add function files and library files to the prerequisites list for the
ZIP target.

Update ZIP target `zip` invocation to use neither the update flag (`-u`)
nor the set last modified time to oldest file in the archive (`-o`).
Previously the ZIP file would always be reproduced due to having a last
modified date older than other files it depends on.

The prerequisites need to be tightened on the ZIP target as right now it
is rebuilding all the ZIP packages when any of the code files change,
rather than the ones belonging to an individual package.

Resolves: #5